# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KanoonAI Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SendOtpRequest {
  phone: string;
  name?: string;
}

export interface SendOtpResponse {
  success: boolean;
  message: string;
  /** OTP for dev mode only */
  devOtp?: string;
}

export interface VerifyOtpRequest {
  phone: string;
  otp: string;
  name?: string;
}

export interface User {
  id: number;
  phone: string;
  name?: string;
  plan: string;
  referralCode?: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserAdmin {
  id: number;
  phone: string;
  name?: string;
  plan: string;
  documentCount: number;
  totalSpent: number;
  createdAt: string;
}

export type DocumentFormData = { [key: string]: unknown };

export interface Document {
  id: number;
  userId: number;
  type: string;
  title: string;
  content?: string;
  formData?: DocumentFormData;
  pdfUrl?: string;
  paid: boolean;
  language: string;
  price: number;
  createdAt: string;
}

export type GenerateDocumentRequestFormData = { [key: string]: unknown };

export type GenerateDocumentRequestLanguage =
  (typeof GenerateDocumentRequestLanguage)[keyof typeof GenerateDocumentRequestLanguage];

export const GenerateDocumentRequestLanguage = {
  en: "en",
  hi: "hi",
  mr: "mr",
  ta: "ta",
  te: "te",
} as const;

export interface GenerateDocumentRequest {
  type: string;
  formData: GenerateDocumentRequestFormData;
  language: GenerateDocumentRequestLanguage;
}

export interface DownloadResponse {
  url: string;
  content?: string;
}

export interface CreateOrderRequest {
  documentId: number;
  amount: number;
}

export interface PaymentOrder {
  orderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface VerifyPaymentRequest {
  orderId: string;
  paymentId: string;
  signature: string;
  documentId: number;
}

export interface PaymentVerifyResponse {
  success: boolean;
  document?: Document;
}

export interface Payment {
  id: number;
  userId?: number;
  documentId?: number;
  amount: number;
  razorpayOrderId?: string;
  razorpayPaymentId?: string;
  status: string;
  plan?: string;
  createdAt: string;
}

export type CreateSubscriptionRequestPlan =
  (typeof CreateSubscriptionRequestPlan)[keyof typeof CreateSubscriptionRequestPlan];

export const CreateSubscriptionRequestPlan = {
  basic: "basic",
  pro: "pro",
  business: "business",
} as const;

export interface CreateSubscriptionRequest {
  plan: CreateSubscriptionRequestPlan;
}

export type AdminStatsPopularDocTypesItem = {
  type: string;
  count: number;
};

export interface AdminStats {
  totalUsers: number;
  totalDocuments: number;
  revenueToday: number;
  revenueMonth: number;
  revenueTotal: number;
  documentsToday: number;
  documentsMonth: number;
  popularDocTypes: AdminStatsPopularDocTypesItem[];
  recentTransactions: Payment[];
}