# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KanoonAI Legal Document Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: documents
    description: Document operations
  - name: payments
    description: Payment operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOtpResponse"

  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List user documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: generateDocument
      tags: [documents]
      summary: Generate a new document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDocumentRequest"
      responses:
        "200":
          description: Generated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Download document PDF
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF download URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadResponse"

  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create Razorpay order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Payment order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"

  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment and unlock document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"

  /payments/history:
    get:
      operationId: getPaymentHistory
      tags: [payments]
      summary: Get payment history for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

  /payments/subscription/create:
    post:
      operationId: createSubscription
      tags: [payments]
      summary: Create subscription order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "200":
          description: Subscription order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAdmin"

  /admin/documents:
    get:
      operationId: listAdminDocuments
      tags: [admin]
      summary: List all documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  /admin/transactions:
    get:
      operationId: listAdminTransactions
      tags: [admin]
      summary: List all transactions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SendOtpRequest:
      type: object
      properties:
        phone:
          type: string
        name:
          type: string
      required:
        - phone

    SendOtpResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        devOtp:
          type: string
          description: OTP for dev mode only
      required:
        - success
        - message

    VerifyOtpRequest:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
        name:
          type: string
      required:
        - phone
        - otp

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        plan:
          type: string
        referralCode:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - phone
        - plan
        - createdAt

    UserAdmin:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        plan:
          type: string
        documentCount:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - phone
        - plan
        - documentCount
        - totalSpent
        - createdAt

    Document:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        content:
          type: string
        formData:
          type: object
          additionalProperties: true
        pdfUrl:
          type: string
        paid:
          type: boolean
        language:
          type: string
        price:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - paid
        - language
        - price
        - createdAt

    GenerateDocumentRequest:
      type: object
      properties:
        type:
          type: string
        formData:
          type: object
          additionalProperties: true
        language:
          type: string
          enum: [en, hi, mr, ta, te]
      required:
        - type
        - formData
        - language

    DownloadResponse:
      type: object
      properties:
        url:
          type: string
        content:
          type: string
      required:
        - url

    CreateOrderRequest:
      type: object
      properties:
        documentId:
          type: integer
        amount:
          type: integer
      required:
        - documentId
        - amount

    PaymentOrder:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
      required:
        - orderId
        - amount
        - currency
        - keyId

    VerifyPaymentRequest:
      type: object
      properties:
        orderId:
          type: string
        paymentId:
          type: string
        signature:
          type: string
        documentId:
          type: integer
      required:
        - orderId
        - paymentId
        - signature
        - documentId

    PaymentVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        document:
          $ref: "#/components/schemas/Document"
      required:
        - success

    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        documentId:
          type: integer
        amount:
          type: integer
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        status:
          type: string
        plan:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - amount
        - status
        - createdAt

    CreateSubscriptionRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [basic, pro, business]
      required:
        - plan

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalDocuments:
          type: integer
        revenueToday:
          type: number
        revenueMonth:
          type: number
        revenueTotal:
          type: number
        documentsToday:
          type: integer
        documentsMonth:
          type: integer
        popularDocTypes:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
            required:
              - type
              - count
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
      required:
        - totalUsers
        - totalDocuments
        - revenueToday
        - revenueMonth
        - revenueTotal
        - documentsToday
        - documentsMonth
        - popularDocTypes
        - recentTransactions