# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KanoonAI Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminStatsPopularDocTypesItem } from "./adminStatsPopularDocTypesItem";
import type { Payment } from "./payment";

export interface AdminStats {
  totalUsers: number;
  totalDocuments: number;
  revenueToday: number;
  revenueMonth: number;
  revenueTotal: number;
  documentsToday: number;
  documentsMonth: number;
  popularDocTypes: AdminStatsPopularDocTypesItem[];
  recentTransactions: Payment[];
}