# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KanoonAI Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentFormData } from "./documentFormData";

export interface Document {
  id: number;
  userId: number;
  type: string;
  title: string;
  content?: string;
  formData?: DocumentFormData;
  pdfUrl?: string;
  paid: boolean;
  language: string;
  price: number;
  createdAt: Date;
}