# lib/api-zod/src/generated/types/generateDocumentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KanoonAI Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateDocumentRequestFormData } from "./generateDocumentRequestFormData";
import type { GenerateDocumentRequestLanguage } from "./generateDocumentRequestLanguage";

export interface GenerateDocumentRequest {
  type: string;
  formData: GenerateDocumentRequestFormData;
  language: GenerateDocumentRequestLanguage;
}