# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KanoonAI Legal Document Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  userId?: number;
  documentId?: number;
  amount: number;
  razorpayOrderId?: string;
  razorpayPaymentId?: string;
  status: string;
  plan?: string;
  createdAt: Date;
}